# src/api/v1/openapi.yaml
openapi: 3.1.0
info:
  title: NiMa Public API
  version: 1.0.0
  description: Public API v1 scaffold for projects, tasks, activity, and analytics overview.
servers:
  - url: http://localhost:8790
paths:
...
  /v1/projects:
    get:
      summary: List projects (placeholder)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Placeholder response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse'
    post:
      summary: Create project (placeholder)
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        '202':
          description: Placeholder accepted response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
...
  /v1/analytics/overview:
    get:
      summary: Workspace-scoped analytics overview
      security:
        - bearerAuth: []
      parameters:
        - name: guildId
          in: query
          required: true
          schema:
            type: string
        - name: workspaceId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Analytics overview response
          content:
            application/json:
              schema:
                type: object
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  schemas:
    Meta:
      type: object
      properties:
        resource:
          type: string
        version:
          type: string
        placeholder:
          type: boolean
        authMode:
          type: string
      required: [resource, version, placeholder, authMode]
    ListResponse:
      type: object
      properties:
        ok:
          type: boolean
        data:
          type: array
          items: {}
        meta:
          $ref: '#/components/schemas/Meta'
      required: [ok, data, meta]
    AcceptedResponse:
      type: object
      properties:
        ok:
          type: boolean
        data:
          type: object
          properties:
            accepted:
              type: boolean
            received:
              type: object
          required: [accepted, received]
        meta:
          $ref: '#/components/schemas/Meta'
      required: [ok, data, meta]